Split Info Dialog into another fragment as well.
